Make reactor/direct2d sample a hub of Direct2D samples
Restructures the single-example direct2d sample into a NavigationView-based hub modeled on the gallery sample. Each sample is a self-contained component selected from the nav pane:

- swap_chain: the existing animated circles presented via SwapChainPanel

- surface_image_source: folded in from the standalone reactor/surface_image_source crate, which is removed
