--- crates/samples/reactor/surface_image_source/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    windows_reactor_setup::as_self_contained();
-}
